requirements.txt
tensorflow==2.20.0
opencv-python
numpy==1.26.4
scikit-learn
streamlit
matplotlib
Pillow
